+++ compiler/bsc/rescript_compiler_main.ml
@@ -50,6 +50,8 @@ let set_abs_input_name sourcefile =
   sourcefile
 let setup_outcome_printer () = Lazy.force Res_outcome_printer.setup
 
+let setup_runtime_path path = Runtime_package.path := path
+
 let process_file sourcefile ?kind ppf =
   (* This is a better default then "", it will be changed later
      The {!Location.input_name} relies on that we write the binary ast
@@ -394,6 +396,9 @@ let command_line_flags : (string * Bsc_args.spec * string) array =
     ( "-unsafe",
       set Clflags.fast,
       "*internal* Do not compile bounds checking on array and string access" );
+    ( "-runtime-path",
+      string_call setup_runtime_path,
+      "*internal* Set the path of the runtime package (@rescript/runtime)" );
     ( "-warn-help",
       unit_call Warnings.help_warnings,
       "Show description of warning numbers" );