+++ compiler/common/bs_version.mli
@@ -25,5 +25,3 @@
 val version : string
 
 val header : string
-
-val package_name : string